requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
nest-asyncio==1.6.0
python-dotenv==1.0.1
requests==2.32.3
upstash-ratelimit==1.1.0
upstash-redis==1.1.0
urllib3==2.2.2
yarl==1.9.4
pandas==2.0.3
numpy==1.24.3
matplotlib
seaborn
scikit-learn
tqdm
tabulate
fpdf
langchain_core
langchain_community
langchain_groq
langchain_huggingface
langchain
itertools
